pyproject.toml
[project]
name = "opensdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.129.2",
    "jinja2>=3.1.6",
    "markdown>=3.10.2",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
